fix Javolution caching for vobs: use Javolution reference counter
When a vob is put into a VobMap, increase that vob's reference counter;
when the vob map is cleared, decrease the reference counter. It can
then be re-used.

Also fix a bug in TableLob, where a Layout object created in a pool context
was referenced after the pool context was exited (i.e., when the object
was allowed to be recycled).

darcs-hash:20050309003436-2ee97-b8f89bdfeabcf65cbb053bb4cc4d571630411b9e