Good architecture for text cursor lob.
This patch introduces a new lob, DecoratorLob. This lob takes a child lob,
a "decoration lob," and a key. It first renders its child, then searches
the vobscene for a cs placed by its child with the given key; then it renders
the decoration into that cs.

For the text cursor, we decorate the cs of the character following the cursor
with a line on the left side.

Unfortunately the decorator has to use getSqSize to be able to layout
the decoration, which is slow on GL. Ideas appreciated.

darcs-hash:20050320161115-2ee97-6af70c56ac1afcc5fd179e3f2e641d11d97e15bf