In TextKeyController, use WindowAnimation for right animation behavior.
(We want to animate when changing the cursor position, but switchVS
when typing.)

Instead of passing the TextKeyController's constructor a WindowAnimation
object, as we would have had to do in the old lob system, we use Javolution's
LocalContext facility. This allows us to create a "thread-local" variable--
like

	    PoolContext.enter();
	    LocalContext.enter();
	    try {
		Lobs.setWindowAnimation(windowAnim);

		if (initialized && createLob().key(key)) {
		    if(!windowAnim.hasSceneReplacementPending())
			windowAnim.animate();
		}
	    } finally {
		LocalContext.exit();
		PoolContext.exit();
	    }

which sets the WindowAnimation only for this block (and only
on the current thread).

darcs-hash:20050321015302-2ee97-295b94f1fc69066aceceeee21bcc68c5e22c6026