A code preprocessor to make Javolution programming a bit easier.
With Javolution, a simple inner class like

    new LobList() {
        public int getLobCount() { list.getLobCount(); }
        public Lob getLob(int index) {
            return KeyLob.newInstance(list.getLob(index));
        }
    }

becomes a really long and complex thing, see KeyLobList for how the above
then looks.

I found this too hard in practice, so with this new preprocessor, this becomes

    new @realtime LobList(LobList list) {
        public int getLobCount() { list.getLobCount(); }
        public Lob getLob(int index) {
            return KeyLob.newInstance(list.getLob(index));
        }
    }

The preprocessor will create an inner class and create a Javolution FACTORY
for that class etc.

The thing after @realtime must be an interface, not a class (for now).

We can not access final members of the enclosing method from the inner class,
though, that was too hard to program, that's why we explicitly have to say

    (LobList list)

in the 'new @realtime' clause.

darcs-hash:20050321171001-2ee97-1f775f1d12afd283cbe64c6b3471c2908decd374